+++ charond/rpc_server_user_permissions.go
@@ -18,10 +18,10 @@ func (rs *rpcServer) SetUserPermissions(ctx context.Context, req *charon.SetUser
 
 // ListUserPermissions implements charon.RPCServer interface.
 func (rs *rpcServer) ListUserPermissions(ctx context.Context, req *charon.ListUserPermissionsRequest) (*charon.ListUserPermissionsResponse, error) {
-	permissions, err := rs.repository.permission.FindByUserID(req.Id)
+	permissions, err := rs.repository.permission.FindByUserID(int64(req.Id))
 	if err != nil {
 		if err == sql.ErrNoRows {
-			sklog.Debug(rs.logger, "user permissions retrieved", "user_id", req.Id, "count", len(permissions))
+			sklog.Debug(rs.logger, "user permissions retrieved", "user_id", int64(req.Id), "count", len(permissions))
 
 			return &charon.ListUserPermissionsResponse{}, nil
 		}
@@ -33,7 +33,7 @@ func (rs *rpcServer) ListUserPermissions(ctx context.Context, req *charon.ListUs
 		perms = append(perms, p.Permission().String())
 	}
 
-	sklog.Debug(rs.logger, "user permissions retrieved", "user_id", req.Id, "count", len(permissions))
+	sklog.Debug(rs.logger, "user permissions retrieved", "user_id", int64(req.Id), "count", len(permissions))
 
 	return &charon.ListUserPermissionsResponse{
 		Permissions: perms,
